mag3110.c write field offset values to chip
The magnetic field offset values due to presence of "hard iron" can be
corrected with the OFF_X, OFF_Y and OFF_Z parameters in the configuration
file. After writing the correction vector on the MAG3110 the values are
read back and they are printed in the log file.

Define TOFFSET in the configuration file. The MAG3110 chip temperature needs
to be corrected by an offset.

Use C shift operators '<<' and '>>' instead of integer division and reminder.
This with more careful type cast should handle the one or two byte negative
integer numbers correctly.